+++ CHANGELOG.md
@@ -7,6 +7,7 @@
 ### Changed
 
 - Add a `python_requires` to enforce support of Python 3.6+ only.
+- Switch from docopt to click, like black, with an open-end version range.
 
 ## [v0.6.0](https://github.com/thibaudcolas/curlylint/releases/tag/v0.6.0) 2020-03-30